Move operations with rollback to Backstore for data and cache device handling
