+++ src/engine/strat_engine/backstore/blockdevmgr.rs
@@ -4,15 +4,12 @@
 
 // Code to handle a collection of block devices.
 
+use std::collections::HashMap;
 #[cfg(test)]
 use std::collections::HashSet;
 
-use std::{collections::HashMap, fs, path::PathBuf};
-
 use chrono::{DateTime, Duration, Utc};
 use rand::{seq::IteratorRandom, thread_rng};
-use serde_json::Value;
-use tempfile::TempDir;
 
 use devicemapper::{Bytes, Device, Sectors};
 
@@ -22,10 +19,6 @@ use crate::{
         strat_engine::{
             backstore::{
                 blockdev::StratBlockDev,
-                crypt::{
-                    back_up_luks_header, interpret_clevis_config, restore_luks_header,
-                    CryptActivationHandle,
-                },
                 devices::{initialize_devices, wipe_blockdevs, UnownedDevices},
                 range_alloc::PerDevSegments,
                 shared::{BlkDevSegment, Segment},
@@ -33,12 +26,9 @@ use crate::{
             },
             metadata::{MDADataSize, BDA},
             serde_structs::{BaseBlockDevSave, Recordable},
-            shared::bds_to_bdas,
+            shared::{bds_to_bdas, can_unlock},
         },
-        types::{
-            ActionAvailability, DevUuid, EncryptionInfo, KeyDescription, Name, PoolEncryptionInfo,
-            PoolUuid,
-        },
+        types::{DevUuid, EncryptionInfo, Name, PoolEncryptionInfo, PoolUuid},
     },
     stratis::{StratisError, StratisResult},
 };
@@ -99,20 +89,6 @@ impl BlockDevMgr {
             .collect()
     }
 
-    /// Check that the registered key description and Clevis information for these
-    /// block devices can unlock at least one of the existing block devices registered.
-    /// Precondition: self.block_devs must have at least one device.
-    pub fn can_unlock(&self, try_unlock_keyring: bool, try_unlock_clevis: bool) -> bool {
-        CryptActivationHandle::can_unlock(
-            self.block_devs
-                .get(0)
-                .expect("Must have at least one blockdev")
-                .physical_path(),
-            try_unlock_keyring,
-            try_unlock_clevis,
-        )
-    }
-
     /// Add paths to self.
     /// Return the uuids of all blockdevs corresponding to paths that were
     /// added.
@@ -133,7 +109,14 @@ impl BlockDevMgr {
 
         let encryption_info = pool_enc_to_enc!(self.encryption_info());
         if let Some(ref ei) = encryption_info {
-            if !self.can_unlock(ei.key_description().is_some(), ei.clevis_info().is_some()) {
+            if !can_unlock(
+                self.block_devs
+                    .get(0)
+                    .expect("Must have at least one blockdev")
+                    .physical_path(),
+                ei.key_description().is_some(),
+                ei.clevis_info().is_some(),
+            ) {
                 return Err(StratisError::Msg(
                     "Neither the key in the kernel keyring nor Clevis could be used to perform encryption operations on the devices in the pool; check that either the appropriate key in the keyring is set or that the Clevis key storage method is available".to_string(),
                 ));
@@ -407,228 +390,6 @@ impl BlockDevMgr {
         }
     }
 
-    /// Bind all devices in the given blockdev manager using the given clevis
-    /// configuration.
-    ///
-    /// * Returns Ok(true) if the binding was performed.
-    /// * Returns Ok(false) if the binding had already been previously performed and
-    /// nothing was changed.
-    /// * Returns Err(_) if an inconsistency was found in the metadata across pools
-    /// or binding failed.
-    pub fn bind_clevis(&mut self, pin: &str, clevis_info: &Value) -> StratisResult<bool> {
-        let encryption_info = match pool_enc_to_enc!(self.encryption_info()) {
-            Some(ei) => ei,
-            None => {
-                return Err(StratisError::Msg(
-                    "Requested pool does not appear to be encrypted".to_string(),
-                ));
-            }
-        };
-
-        let mut parsed_config = clevis_info.clone();
-        let yes = interpret_clevis_config(pin, &mut parsed_config)?;
-
-        if let Some((ref existing_pin, ref existing_info)) = encryption_info.clevis_info() {
-            // Ignore thumbprint if stratis:tang:trust_url is set in the clevis_info
-            // config.
-            let mut config_to_check = existing_info.clone();
-            if yes {
-                if let Value::Object(ref mut ei) = config_to_check {
-                    ei.remove("thp");
-                }
-            }
-
-            if (existing_pin.as_str(), &config_to_check) == (pin, &parsed_config)
-                && self.can_unlock(false, true)
-            {
-                Ok(false)
-            } else {
-                Err(StratisError::Msg(format!(
-                    "Block devices have already been bound with pin {} and config {}; \
-                        requested pin {} and config {} can't be applied",
-                    existing_pin, existing_info, pin, parsed_config,
-                )))
-            }
-        } else {
-            operation_loop(
-                self.blockdevs_mut().into_iter().map(|(_, bd)| bd),
-                |blockdev| blockdev.bind_clevis(pin, clevis_info),
-            )?;
-            Ok(true)
-        }
-    }
-
-    /// Unbind all devices in the given blockdev manager from clevis.
-    ///
-    /// * Returns Ok(true) if the unbinding was performed.
-    /// * Returns Ok(false) if the unbinding had already been previously performed and
-    /// nothing was changed.
-    /// * Returns Err(_) if an inconsistency was found in the metadata across pools
-    /// or unbinding failed.
-    pub fn unbind_clevis(&mut self) -> StratisResult<bool> {
-        let encryption_info = match pool_enc_to_enc!(self.encryption_info()) {
-            Some(ei) => ei,
-            None => {
-                return Err(StratisError::Msg(
-                    "Requested pool does not appear to be encrypted".to_string(),
-                ));
-            }
-        };
-
-        if encryption_info.clevis_info().is_some() {
-            operation_loop(
-                self.blockdevs_mut().into_iter().map(|(_, bd)| bd),
-                |blockdev| blockdev.unbind_clevis(),
-            )?;
-            Ok(true)
-        } else {
-            Ok(false)
-        }
-    }
-
-    /// Bind all devices in the given blockdev manager to a passphrase using the
-    /// given key description.
-    ///
-    /// * Returns Ok(true) if the binding was performed.
-    /// * Returns Ok(false) if the binding had already been previously performed and
-    /// nothing was changed.
-    /// * Returns Err(_) if an inconsistency was found in the metadata across pools
-    /// or binding failed.
-    pub fn bind_keyring(&mut self, key_desc: &KeyDescription) -> StratisResult<bool> {
-        let encryption_info = match pool_enc_to_enc!(self.encryption_info()) {
-            Some(ei) => ei,
-            None => {
-                return Err(StratisError::Msg(
-                    "Requested pool does not appear to be encrypted".to_string(),
-                ));
-            }
-        };
-
-        if let Some(kd) = encryption_info.key_description() {
-            if kd == key_desc {
-                if self.can_unlock(true, false) {
-                    Ok(false)
-                } else {
-                    Err(StratisError::Msg(format!(
-                        "Key description {} is registered in the metadata but the \
-                            associated passphrase can't unlock the device; the \
-                            associated passphrase may have changed since binding",
-                        key_desc.as_application_str(),
-                    )))
-                }
-            } else {
-                Err(StratisError::Msg(format!(
-                    "Block devices have already been bound with key description {}; \
-                        requested key description {} can't be applied",
-                    key_desc.as_application_str(),
-                    kd.as_application_str(),
-                )))
-            }
-        } else {
-            operation_loop(
-                self.blockdevs_mut().into_iter().map(|(_, bd)| bd),
-                |blockdev| blockdev.bind_keyring(key_desc),
-            )?;
-            Ok(true)
-        }
-    }
-
-    /// Unbind all devices in the given blockdev manager from the passphrase
-    /// associated with the key description.
-    ///
-    /// * Returns Ok(true) if the unbinding was performed.
-    /// * Returns Ok(false) if the unbinding had already been previously performed and
-    /// nothing was changed.
-    /// * Returns Err(_) if an inconsistency was found in the metadata across pools
-    /// or unbinding failed.
-    pub fn unbind_keyring(&mut self) -> StratisResult<bool> {
-        let encryption_info = match pool_enc_to_enc!(self.encryption_info()) {
-            Some(ei) => ei,
-            None => {
-                return Err(StratisError::Msg(
-                    "Requested pool does not appear to be encrypted".to_string(),
-                ));
-            }
-        };
-
-        if encryption_info.key_description().is_some() {
-            operation_loop(
-                self.blockdevs_mut().into_iter().map(|(_, bd)| bd),
-                |blockdev| blockdev.unbind_keyring(),
-            )?;
-            Ok(true)
-        } else {
-            // is encrypted and key description is None
-            Ok(false)
-        }
-    }
-
-    /// Change the keyring passphrase associated with the block devices in
-    /// this pool.
-    ///
-    /// Returns:
-    /// * Ok(None) if the pool is not currently bound to a keyring passphrase.
-    /// * Ok(Some(true)) if the pool was successfully bound to the new key description.
-    /// * Ok(Some(false)) if the pool is already bound to this key description.
-    /// * Err(_) if an operation fails while changing the passphrase.
-    pub fn rebind_keyring(&mut self, key_desc: &KeyDescription) -> StratisResult<Option<bool>> {
-        let encryption_info = match pool_enc_to_enc!(self.encryption_info()) {
-            Some(ei) => ei,
-            None => {
-                return Err(StratisError::Msg(
-                    "Requested pool does not appear to be encrypted".to_string(),
-                ));
-            }
-        };
-
-        if encryption_info.key_description() == Some(key_desc) {
-            Ok(Some(false))
-        } else if encryption_info.key_description().is_some() {
-            // Keys are not the same but key description is present
-            operation_loop(
-                self.blockdevs_mut().into_iter().map(|(_, bd)| bd),
-                |blockdev| blockdev.rebind_keyring(key_desc),
-            )?;
-            Ok(Some(true))
-        } else {
-            Ok(None)
-        }
-    }
-
-    /// Regenerate the Clevis bindings with the block devices in this pool using
-    /// the same configuration.
-    ///
-    /// The method for this rollback caches the initial Clevis metadata and
-    /// reverts all of the devices if there is a failure.
-    ///
-    /// This method returns StratisResult<()> because the Clevis regen command
-    /// will always change the metadata when successful. The command is not idempotent
-    /// so this method will either fail to regenerate the bindings or it will
-    /// result in a metadata change.
-    pub fn rebind_clevis(&mut self) -> StratisResult<()> {
-        let encryption_info = match pool_enc_to_enc!(self.encryption_info()) {
-            Some(ei) => ei,
-            None => {
-                return Err(StratisError::Msg(
-                    "Requested pool does not appear to be encrypted".to_string(),
-                ));
-            }
-        };
-
-        if encryption_info.clevis_info().is_none() {
-            Err(StratisError::Msg(
-                "Requested pool is not already bound to Clevis".to_string(),
-            ))
-        } else {
-            operation_loop(
-                self.blockdevs_mut().into_iter().map(|(_, bd)| bd),
-                |blockdev| blockdev.rebind_clevis(),
-            )?;
-
-            Ok(())
-        }
-    }
-
     pub fn grow(&mut self, dev: DevUuid) -> StratisResult<bool> {
         let bd = self
             .block_devs
@@ -638,17 +399,6 @@ impl BlockDevMgr {
                 StratisError::Msg(format!("Block device with UUID {} not found", dev))
             })?;
         bd.grow()
-    }
-
-    /// Rename pool name in LUKS2 token if pool is encrypted.
-    pub fn rename_pool(&mut self, new_name: &Name) -> StratisResult<()> {
-        if self.is_encrypted() {
-            operation_loop(
-                self.blockdevs_mut().into_iter().map(|(_, bd)| bd),
-                |blockdev| blockdev.rename_pool(new_name.clone()),
-            )?;
-        }
-        Ok(())
     }
 
     /// Tear down devicemapper devices for the block devices in this BlockDevMgr.
@@ -679,72 +429,6 @@ impl BlockDevMgr {
     }
 }
 
-fn operation_loop<'a, I, A>(blockdevs: I, action: A) -> StratisResult<()>
-where
-    I: IntoIterator<Item = &'a mut StratBlockDev>,
-    A: Fn(&mut StratBlockDev) -> StratisResult<()>,
-{
-    fn rollback_loop(
-        rollback_record: Vec<&mut StratBlockDev>,
-        headers: Vec<PathBuf>,
-        causal_error: StratisError,
-    ) -> StratisError {
-        // NOTE: Zip can be used here because the header will always be backed up before
-        // the operation is performed. As a result, the header iterator will always be
-        // equal to or longer than the blockdev record iterator which means all blockdevs
-        // that have had operations performed on them will always be restored.
-        for (blockdev, header) in rollback_record.into_iter().zip(headers) {
-            if let Err(e) = restore_luks_header(blockdev.devnode(), header.as_path()) {
-                warn!(
-                    "Failed to roll back device operation for device {}: {}",
-                    blockdev.physical_path().display(),
-                    e,
-                );
-                return StratisError::RollbackError {
-                    causal_error: Box::new(causal_error),
-                    rollback_error: Box::new(e),
-                    level: ActionAvailability::NoRequests,
-                };
-            }
-        }
-
-        causal_error
-    }
-
-    fn perform_operation<'a, I, A>(tmp_dir: &TempDir, blockdevs: I, action: A) -> StratisResult<()>
-    where
-        I: IntoIterator<Item = &'a mut StratBlockDev>,
-        A: Fn(&mut StratBlockDev) -> StratisResult<()>,
-    {
-        let mut original_headers = Vec::new();
-        let mut rollback_record = Vec::new();
-        for blockdev in blockdevs {
-            match back_up_luks_header(blockdev.physical_path(), tmp_dir) {
-                Ok(h) => original_headers.push(h),
-                Err(e) => return Err(rollback_loop(rollback_record, original_headers, e)),
-            };
-            let res = action(blockdev);
-            rollback_record.push(blockdev);
-            if let Err(error) = res {
-                return Err(rollback_loop(rollback_record, original_headers, error));
-            }
-        }
-
-        Ok(())
-    }
-
-    let tmp_dir = TempDir::new()?;
-    let res = perform_operation(&tmp_dir, blockdevs, action);
-    if let Err(e) = fs::remove_dir_all(tmp_dir.path()) {
-        warn!(
-            "Leaked temporary files at path {}: {}",
-            tmp_dir.path().display(),
-            e
-        );
-    }
-    res
-}
-
 impl Recordable<Vec<BaseBlockDevSave>> for BlockDevMgr {
     fn record(&self) -> Vec<BaseBlockDevSave> {
         self.block_devs.iter().map(|bd| bd.record()).collect()
@@ -753,14 +437,15 @@ impl Recordable<Vec<BaseBlockDevSave>> for BlockDevMgr {
 
 #[cfg(test)]
 mod tests {
-    use std::{env, error::Error, path::Path};
+    use std::{error::Error, path::Path};
 
-    use crate::engine::strat_engine::{
-        backstore::devices::{ProcessedPathInfos, UnownedDevices},
-        cmd,
-        names::KeyDescription,
-        ns::{unshare_mount_namespace, MemoryFilesystem},
-        tests::{crypt, loopbacked, real},
+    use crate::engine::{
+        strat_engine::{
+            backstore::devices::{ProcessedPathInfos, UnownedDevices},
+            cmd,
+            tests::{crypt, loopbacked, real},
+        },
+        types::KeyDescription,
     };
 
     use super::*;
@@ -982,188 +667,6 @@ mod tests {
         real::test_with_spec(
             &real::DeviceLimits::AtLeast(2, None, None),
             test_initialization_add_stratis,
-        );
-    }
-
-    fn test_clevis_initialize(paths: &[&Path]) {
-        unshare_mount_namespace().unwrap();
-        let pool_name = Name::new("pool_name".to_string());
-        let _memfs = MemoryFilesystem::new().unwrap();
-        let pool_uuid = PoolUuid::new_v4();
-        let mut mgr = BlockDevMgr::initialize(
-            pool_name,
-            pool_uuid,
-            get_devices(paths).unwrap(),
-            MDADataSize::default(),
-            Some(&EncryptionInfo::ClevisInfo((
-                "tang".to_string(),
-                json!({"url": env::var("TANG_URL").unwrap(), "stratis:tang:trust_url": true}),
-            ))),
-        )
-        .unwrap();
-        cmd::udev_settle().unwrap();
-
-        matches!(
-            mgr.bind_clevis(
-                "tang",
-                &json!({"url": env::var("TANG_URL").unwrap(), "stratis:tang:trust_url": true})
-            ),
-            Ok(false)
-        );
-    }
-
-    #[test]
-    fn clevis_real_test_initialize() {
-        real::test_with_spec(
-            &real::DeviceLimits::AtLeast(2, None, None),
-            test_clevis_initialize,
-        );
-    }
-
-    #[test]
-    #[should_panic]
-    fn clevis_real_should_fail_test_initialize() {
-        real::test_with_spec(
-            &real::DeviceLimits::AtLeast(2, None, None),
-            test_clevis_initialize,
-        );
-    }
-
-    #[test]
-    fn clevis_loop_test_initialize() {
-        loopbacked::test_with_spec(
-            &loopbacked::DeviceLimits::Range(2, 4, None),
-            test_clevis_initialize,
-        );
-    }
-
-    #[test]
-    #[should_panic]
-    fn clevis_loop_should_fail_test_initialize() {
-        loopbacked::test_with_spec(
-            &loopbacked::DeviceLimits::Range(2, 4, None),
-            test_clevis_initialize,
-        );
-    }
-
-    fn test_clevis_both_initialize(paths: &[&Path]) {
-        fn test_both(paths: &[&Path], key_desc: &KeyDescription) -> Result<(), Box<dyn Error>> {
-            unshare_mount_namespace()?;
-            let _memfs = MemoryFilesystem::new().unwrap();
-            let pool_uuid = PoolUuid::new_v4();
-            let pool_name = Name::new("pool_name".to_string());
-            let mut mgr = BlockDevMgr::initialize(
-                pool_name,
-                pool_uuid,
-                get_devices(paths).unwrap(),
-                MDADataSize::default(),
-                Some(&EncryptionInfo::Both(
-                    key_desc.clone(),
-                    (
-                        "tang".to_string(),
-                        json!({"url": env::var("TANG_URL")?, "stratis:tang:trust_url": true}),
-                    ),
-                )),
-            )?;
-            cmd::udev_settle()?;
-
-            if mgr.bind_clevis(
-                "tang",
-                &json!({"url": env::var("TANG_URL")?, "stratis:tang:trust_url": true}),
-            )? {
-                return Err(Box::new(StratisError::Msg(
-                    "Clevis bind idempotence test failed".to_string(),
-                )));
-            }
-            if mgr.bind_keyring(key_desc)? {
-                return Err(Box::new(StratisError::Msg(
-                    "Keyring bind idempotence test failed".to_string(),
-                )));
-            }
-
-            if !(mgr.unbind_clevis()?) {
-                return Err(Box::new(StratisError::Msg(
-                    "Clevis unbind test failed".to_string(),
-                )));
-            }
-            if mgr.unbind_clevis()? {
-                return Err(Box::new(StratisError::Msg(
-                    "Clevis unbind idempotence test failed".to_string(),
-                )));
-            }
-            if mgr.unbind_keyring().is_ok() {
-                return Err(Box::new(StratisError::Msg(
-                    "Keyring unbind check test failed".to_string(),
-                )));
-            }
-
-            if !(mgr.bind_clevis(
-                "tang",
-                &json!({"url": env::var("TANG_URL")?, "stratis:tang:trust_url": true}),
-            )?) {
-                return Err(Box::new(StratisError::Msg(
-                    "Clevis bind test failed".to_string(),
-                )));
-            }
-            if !(mgr.unbind_keyring()?) {
-                return Err(Box::new(StratisError::Msg(
-                    "Keyring unbind test failed".to_string(),
-                )));
-            }
-            if mgr.unbind_keyring()? {
-                return Err(Box::new(StratisError::Msg(
-                    "Keyring unbind idempotence test failed".to_string(),
-                )));
-            }
-            if mgr.unbind_clevis().is_ok() {
-                return Err(Box::new(StratisError::Msg(
-                    "Clevis unbind check test failed".to_string(),
-                )));
-            }
-
-            if !(mgr.bind_keyring(key_desc)?) {
-                return Err(Box::new(StratisError::Msg(
-                    "Keyring bind test failed".to_string(),
-                )));
-            }
-
-            Ok(())
-        }
-
-        crypt::insert_and_cleanup_key(paths, test_both);
-    }
-
-    #[test]
-    fn clevis_real_test_both_initialize() {
-        real::test_with_spec(
-            &real::DeviceLimits::AtLeast(2, None, None),
-            test_clevis_both_initialize,
-        );
-    }
-
-    #[test]
-    #[should_panic]
-    fn clevis_real_should_fail_test_both_initialize() {
-        real::test_with_spec(
-            &real::DeviceLimits::AtLeast(2, None, None),
-            test_clevis_both_initialize,
-        );
-    }
-
-    #[test]
-    fn clevis_loop_test_both_initialize() {
-        loopbacked::test_with_spec(
-            &loopbacked::DeviceLimits::Range(2, 4, None),
-            test_clevis_both_initialize,
-        );
-    }
-
-    #[test]
-    #[should_panic]
-    fn clevis_loop_should_fail_test_both_initialize() {
-        loopbacked::test_with_spec(
-            &loopbacked::DeviceLimits::Range(2, 4, None),
-            test_clevis_both_initialize,
         );
     }
 }
