+++ src/engine/strat_engine/shared.rs
@@ -2,10 +2,13 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-use std::collections::HashMap;
+use std::{collections::HashMap, path::Path};
 
 use crate::engine::{
-    strat_engine::{backstore::StratBlockDev, metadata::BDA},
+    strat_engine::{
+        backstore::{CryptActivationHandle, StratBlockDev},
+        metadata::BDA,
+    },
     types::DevUuid,
 };
 
@@ -28,4 +31,11 @@ pub fn tiers_to_bdas(
         .chain(bds_to_bdas(cachedevs))
         .chain(bda.map(|bda| (bda.dev_uuid(), bda)))
         .collect::<HashMap<_, _>>()
+}
+
+/// Check that the registered key description and Clevis information for these
+/// block devices can unlock at least one of the existing block devices registered.
+/// Precondition: self.block_devs must have at least one device.
+pub fn can_unlock(physical_path: &Path, try_unlock_keyring: bool, try_unlock_clevis: bool) -> bool {
+    CryptActivationHandle::can_unlock(physical_path, try_unlock_keyring, try_unlock_clevis)
 }
